convert libibl from double to float
since libibl is now used online, it makes more sense to do all the
math in float instead of double. It doesn't seem to impact quality.

+++ libs/ibl/include/ibl/Cubemap.h
@@ -84,27 +84,27 @@ public:
     inline Image& getImageForFace(Face face);
 
     //! computes the center of a pixel at coordinate x, y
-    inline filament::math::double2 center(size_t x, size_t y) const;
+    inline filament::math::float2 center(size_t x, size_t y) const;
 
     //! computes a direction vector from a face and a location of the center of pixel in an Image
-    inline filament::math::double3 getDirectionFor(Face face, size_t x, size_t y) const;
+    inline filament::math::float3 getDirectionFor(Face face, size_t x, size_t y) const;
 
     //! computes a direction vector from a face and a location in pixel in an Image
-    inline filament::math::double3 getDirectionFor(Face face, double x, double y) const;
+    inline filament::math::float3 getDirectionFor(Face face, float x, float y) const;
 
     //! samples the cubemap at the given direction using nearest neighbor filtering
-    inline Texel const& sampleAt(const filament::math::double3& direction) const;
+    inline Texel const& sampleAt(const filament::math::float3& direction) const;
 
     //! samples the cubemap at the given direction using bilinear filtering
-    inline Texel filterAt(const filament::math::double3& direction) const;
+    inline Texel filterAt(const filament::math::float3& direction) const;
 
     //! samples an image at the given location in pixel using bilinear filtering
     static Texel filterAt(const Image& image, float x, float y);
     static Texel filterAtCenter(const Image& image, size_t x, size_t y);
 
     //! samples two cubemaps in a given direction and lerps the result by a given lerp factor
-    static Texel trilinearFilterAt(const Cubemap& c0, const Cubemap& c1, double lerp,
-            const filament::math::double3& direction);
+    static Texel trilinearFilterAt(const Cubemap& c0, const Cubemap& c1, float lerp,
+            const filament::math::float3& direction);
 
     //! reads a texel at a given address
     inline static const Texel& sampleAt(void const* data) {
@@ -129,17 +129,17 @@ public:
 
     struct Address {
         Face face;
-        double s = 0;
-        double t = 0;
+        float s = 0;
+        float t = 0;
     };
 
     //! returns the face and texture coordinates of the given direction
-    static Address getAddressFor(const filament::math::double3& direction);
+    static Address getAddressFor(const filament::math::float3& direction);
 
 private:
     size_t mDimensions = 0;
-    double mScale = 1;
-    double mUpperBound = 0;
+    float mScale = 1;
+    float mUpperBound = 0;
     Image mFaces[6];
 };
 
@@ -153,21 +153,21 @@ inline Image& Cubemap::getImageForFace(Face face) {
     return mFaces[int(face)];
 }
 
-inline filament::math::double2 Cubemap::center(size_t x, size_t y) const {
-    return {x + 0.5, y + 0.5};
+inline filament::math::float2 Cubemap::center(size_t x, size_t y) const {
+    return { x + 0.5f, y + 0.5f };
 }
 
-inline filament::math::double3 Cubemap::getDirectionFor(Face face, size_t x, size_t y) const {
-    return getDirectionFor(face, x + 0.5, y + 0.5);
+inline filament::math::float3 Cubemap::getDirectionFor(Face face, size_t x, size_t y) const {
+    return getDirectionFor(face, x + 0.5f, y + 0.5f);
 }
 
-inline filament::math::double3 Cubemap::getDirectionFor(Face face, double x, double y) const {
+inline filament::math::float3 Cubemap::getDirectionFor(Face face, float x, float y) const {
     // map [0, dim] to [-1,1] with (-1,-1) at bottom left
-    double cx = (x * mScale) - 1;
-    double cy = 1 - (y * mScale);
+    float cx = (x * mScale) - 1;
+    float cy = 1 - (y * mScale);
 
-    filament::math::double3 dir;
-    const double l = std::sqrt(cx * cx + cy * cy + 1);
+    filament::math::float3 dir;
+    const float l = std::sqrt(cx * cx + cy * cy + 1);
     switch (face) {
         case Face::PX:  dir = {   1, cy, -cx }; break;
         case Face::NX:  dir = {  -1, cy,  cx }; break;
@@ -179,14 +179,14 @@ inline filament::math::double3 Cubemap::getDirectionFor(Face face, double x, dou
     return dir * (1 / l);
 }
 
-inline Cubemap::Texel const& Cubemap::sampleAt(const filament::math::double3& direction) const {
+inline Cubemap::Texel const& Cubemap::sampleAt(const filament::math::float3& direction) const {
     Cubemap::Address addr(getAddressFor(direction));
     const size_t x = std::min(size_t(addr.s * mDimensions), mDimensions - 1);
     const size_t y = std::min(size_t(addr.t * mDimensions), mDimensions - 1);
     return sampleAt(getImageForFace(addr.face).getPixelRef(x, y));
 }
 
-inline Cubemap::Texel Cubemap::filterAt(const filament::math::double3& direction) const {
+inline Cubemap::Texel Cubemap::filterAt(const filament::math::float3& direction) const {
     Cubemap::Address addr(getAddressFor(direction));
     addr.s = std::min(addr.s * mDimensions, mUpperBound);
     addr.t = std::min(addr.t * mDimensions, mUpperBound);
